embeddings/requirements.txt
fastapi
uvicorn
open_clip_torch
torch
Pillow
numpy
python-multipart

# test
pytest
pytest-asyncio
httpx